"Bump version and update generic type parameters"
This commit includes two changes:

Firstly, the version in magisterapi/build.gradle.kts was bumped from 1.1.10-alpha09
to 1.1.10-alpha10 to signify a new iteration in the development process.

Secondly, in the MessageFlow.kt file, the generic parameters for the data classes
PatchMessageRequest, PatchMessage, and PatchOperation were updated from read-only
(out T) to mutable (T). This was done to allow modifications to the message
objects after instantiation, providing more flexibility in handling messages.