Add Serialization Strategy for Patch Requests
In the MessageFlow file, the method of creating patch requests has been updated. Previously, the requestPATCH function was being used with individual arguments. However, for better management and handling of patch requests, the SerializationStrategy has been added along with the encodeToString function. This will not only improve the readability of the code but also make error tracking easier. Moreover, using the SerializationStrategy provides a framework for seralizing objects, making the process more efficient.