Refactor endpoint configuration and conditional signing checks
Updated the LoginFlow.kt in the magisterapi by refactoring how we configure the main, authorization, and token endpoints. Introduced setters for these Urls for better flexibility and modularity. The Urls are now built using URLBuilder, which takes the main endpoint as a base and appends encoded path segments. This change will allow us to dynamically modify the endpoints if needed.

Also added a conditional check in convention.publication.gradle.kts to see whether signing keys are provided before attempting to sign artifacts. This prevents a possible error when the required signing keys are not provided.Refactor URL configurations and signing checks for flexibility

Updated the URL configurations in the magisterapi and added conditional checks for signing keys. The update includes setters for main, authorization and token endpoints, and the incorporation of URLBuilder for better modularity and flexibility. A conditional check has been added in the convention.publication.gradle.kts to ensure signing keys are provided before signing artifacts, preventing possible errors.